completely abstracted out all of the logic for keystores and separated it from UI components like useEffect
