+++ www/pages/api/uploadJson.ts
@@ -1,5 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { create } from 'ipfs-http-client';
+import { CID, create } from 'ipfs-http-client';
 import { AbortController } from 'node-abort-controller';
 
 global.AbortController = AbortController;
@@ -12,12 +12,8 @@ type Body = {
   data: string;
 };
 
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<Response>
-) {
-  const body = req.body as Body;
-
+export async function uploadJson(body: any): Promise<string> {
+  console.log('body:', body);
   const projectId = process.env.INFURA_IPFS_PROJECT_ID;
   const projectSecret = process.env.INFURA_IPFS_PROJECT_SECRET;
   const auth =
@@ -50,6 +46,13 @@ export default async function handler(
   });
   const graphIpfsResult = await graphIpfs.add(input);
   console.log('uploaded to the graph ipfs node {}', graphIpfsResult);
+  return addResult.cid.toString();
+}
 
-  res.status(200).json({ cid: addResult.cid.toString() });
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse<Response>
+) {
+  const body = req.body as Body;
+  res.status(200).json({ cid: await uploadJson(body) });
 }
